third_party/WebKit/LayoutTests/virtual/color_space/fast/canvas/color-space/canvas-colorManaged-toBlob.html

<!DOCTYPE HTML>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script>

// The reference values are generated by calling SkImage::makeColorSpace()
// in a Skia fiddle as this is the API used in ImageBitmap for color conversion.
// SkColorSpaceXform() may generate slightly different results.
// Please see: https://fiddle.skia.org/c/94578944d9d52c2b4c2cadda88a4e204

function testPixels(ctx1, ctx2, tests)
{
    var actual, expected, tolerance = 0.05;
    for (var i = 0; i < tests.length; i++) {
        actual = ctx1.getImageData(tests[i][0], tests[i][1], 1, 1).dataUnion;
        expected = ctx2.getImageData(tests[i][0], tests[i][1], 1, 1).dataUnion;
        assert_true(actual.length === expected.length);
       for (var i = 0; i < actual.length; i++)
           assert_approx_equals(actual[i], expected[i], tolerance, tests[i][2]);
    }
}


function generateFillStyle(red, green, blue, alpha) {
    return "rgba(" + red + "," + green + "," + blue + "," + alpha + ")";
}

function createTestCanvasAndImageData(testOptions)
{
    var canvas = document.createElement("canvas");
    canvas.width = 2;
    canvas.height = 2;
    var ctx = canvas.getContext('2d', {colorSpace: testOptions.colorSpace, 
                                       pixelFormat: testOptions.pixelFormat});
    ctx.fillStyle = generateFillStyle(155, 27, 27, testOptions.alpha);
    ctx.fillRect(0, 0, 1, 1);
    ctx.fillStyle = generateFillStyle(27, 155, 27, testOptions.alpha);
    ctx.fillRect(1, 0, 1, 1);
    ctx.fillStyle = generateFillStyle(27, 27, 155, testOptions.alpha);
    ctx.fillRect(0, 1, 1, 1);
    ctx.fillStyle = generateFillStyle(27, 27, 27, testOptions.alpha);
    ctx.fillRect(1, 1, 1, 1);
    imageData = ctx.getImageData(0, 0, 2, 2).dataUnion;
    var result = {};
    result.canvas = canvas;
    result.imageData = imageData;
    return result;
}

function printTestOptions(testOptions) {
    console.log ("testOptions: " + testOptions.colorSpace + ", " +
        testOptions.pixelFormat + ", " + testOptions.alpha);
}

function runTest(testOptions) {
    var initVars = createTestCanvasAndImageData(testOptions);
    var srcCanvas = initVars.canvas;
    var srcData = initVars.imageData;
    printTestOptions(testOptions);
    async_test(t => {
        srcCanvas.toBlob(function(blob) {
            createImageBitmap(blob).then(function(bitmap) {
                var dstCanvas = document.createElement("canvas");
                dstCanvas.width = 2;
                dstCanvas.height = 2;
                var ctx = dstCanvas.getContext('2d', 
                        {colorSpace: testOptions.colorSpace, 
                         pixelFormat:testOptions.pixelFormat});
                ctx.drawImage(bitmap, 0, 0);
                var dstData = ctx.getImageData(0, 0, 2, 2).dataUnion;
                printTestOptions(testOptions);
                console.log("srcData: " + srcData);
                console.log("dstData: " + dstData);
                t.done();
            });
        }, "image/webp");
    }, "Test if toBlob creates a Blob-type object.");
}

function runAllTests() {
    var testOptionsSet = [
        // {colorSpace: 'srgb', pixelFormat: '8-8-8-8', alpha: 1},
        // {colorSpace: 'srgb', pixelFormat: 'float16', alpha: 1},
        // {colorSpace: 'rec2020', pixelFormat: 'float16', alpha: 1},
        // {colorSpace: 'p3', pixelFormat: 'float16', alpha: 1},

        // {colorSpace: 'srgb', pixelFormat: '8-8-8-8', alpha: 0.5},
        {colorSpace: 'srgb', pixelFormat: 'float16', alpha: 0.5},
        // {colorSpace: 'rec2020', pixelFormat: 'float16', alpha: 0.5},
        // {colorSpace: 'p3', pixelFormat: 'float16', alpha: 0.5}
    ];

    for (var i = 0; i < testOptionsSet.length; i++)
        runTest(testOptionsSet[i]);
}

runAllTests();

</script>
